requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
python-telegram-bot==13.15
google-cloud-pubsub==2.23.0
google-cloud-firestore==2.16.1
google-cloud-secret-manager==2.20.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pytest==8.3.2
requests-mock==1.12.1
